Create app organization and relocate appropriate objects there
