+++ flash/organization/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class OrganizationConfig(AppConfig):
+    name = 'flash.organization'
