+++ flash/product/admin.py
@@ -1,15 +1,6 @@
 from django.contrib import admin
-from flash.product.models import Product, Category, Organization, Filial
 
-
-@admin.register(Organization)
-class OrganizationAdmin(admin.ModelAdmin):
-    list_display = ('id', 'name', 'rating', 'manager',)
-
-
-@admin.register(Filial)
-class FilialAdmin(admin.ModelAdmin):
-    list_display = ('id', 'address', 'organization',)
+from flash.product.models import Product, Category
 
 
 @admin.register(Category)
